+++ hydromt_wflow/wflow.py
@@ -243,7 +243,7 @@ class WflowModel(Model):
         river_upa=30,
         rivdph_method="powlaw",
         slope_len=2e3,
-        min_rivlen_ratio=0.1,
+        min_rivlen_ratio=0.0,
         min_rivdph=1,
         min_rivwth=30,
         smooth_len=5e3,
@@ -257,7 +257,9 @@ class WflowModel(Model):
         `river_upa` [km2].
 
         The river length is defined as the distance from the subgrid outlet pixel to
-        the next upstream subgrid outlet pixel.
+        the next upstream subgrid outlet pixel. The `min_rivlen_ratio` is the minimum
+        global river length to avg. cell resolution ratio and is used as a threshold in
+        window based smoothing of river length.
 
         The river slope is derived from the subgrid elevation difference between pixels at a
         half distance `slope_len` [m] up- and downstream from the subgrid outlet pixel.
@@ -302,7 +304,10 @@ class WflowModel(Model):
         slope_len : float
             length over which the river slope is calculated [km]
         min_rivlen_ratio: float
-            minimum global river length to avg. cell resolution ratio, by default 0.1
+            Ratio of cell resolution used minimum length threshold in a moving
+            window based smoothing of river length, by default 0.0
+            The river length smoothing is skipped if min_riverlen_ratio = 0.
+            For details about the river length smoothing, see :py:meth:`pyflwdir.FlwdirRaster.smooth_rivlen`
         rivdph_method : {'gvf', 'manning', 'powlaw'}
             see py:meth:`hydromt.workflows.river_depth` for details, by default "powlaw"
         smooth_len : float, optional
@@ -645,17 +650,17 @@ class WflowModel(Model):
     ):
         """
         This component derives several wflow maps are derived based on landuse-
-        landcover (LULC) data. 
-        
+        landcover (LULC) data.
+
         Currently, ``lulc_fn`` can be set to the "vito", "globcover"
         or "corine", fo which lookup tables are constructed to convert lulc classses to
         model parameters based on literature. The data is remapped at its original
         resolution and then resampled to the model resolution using the average
         value, unless noted differently.
-        
+
         Adds model layers:
 
-        * **landuse** map: Landuse class [-]     
+        * **landuse** map: Landuse class [-]
         * **Kext** map: Extinction coefficient in the canopy gap fraction equation [-]
         * **Sl** map: Specific leaf storage [mm]
         * **Swood** map: Fraction of wood in the vegetation/plant [-]
@@ -697,7 +702,7 @@ class WflowModel(Model):
         rmdict = {k: v for k, v in self._MAPS.items() if k in ds_lulc_maps.data_vars}
         self.set_staticmaps(ds_lulc_maps.rename(rmdict))
 
-    def setup_laimaps(self, lai_fn="model_lai"):
+    def setup_laimaps(self, lai_fn="modis_lai"):
         """
         This component sets leaf area index (LAI) climatology maps per month.
 
@@ -1439,6 +1444,84 @@ class WflowModel(Model):
             da_param = da_param.rename(key)
             self.set_staticmaps(da_param)
 
+    def setup_staticmaps_from_raster(
+        self,
+        raster_fn: str,
+        reproject_method: str,
+        variables: Optional[List] = None,
+        wflow_variables: Optional[List] = None,
+        fill_method: Optional[str] = None,
+    ) -> List[str]:
+        """
+        This component adds data variable(s) from ``raster_fn`` to staticmaps object.
+        If raster is a dataset, all variables will be added unless ``variables`` list is specified.
+        The config toml can also be updated to include the new maps using ``wflow_variables``.
+
+        Adds model layers:
+        * **raster.name** or **variables** staticmaps: data from raster_fn
+
+        Parameters
+        ----------
+        raster_fn: str
+            Source name of raster data in data_catalog.
+        reproject_method: str
+            Reprojection method from rasterio.enums.Resampling.
+            Available methods: ['nearest', 'bilinear', 'cubic', 'cubic_spline', 'lanczos', 'average', 'mode',
+            'gauss', 'max', 'min', 'med', 'q1', 'q3', 'sum', 'rms']
+        variables: list, optional
+            List of variables to add to staticmaps from raster_fn. By default all.
+        wflow_variables: list, optional
+            List of corresponding wflow variables to update the config toml (e.g: ["input.vertical.altitude"]).
+            Should match the variables list. variables list should be provided unless raster_fn contains
+            a single variable (len 1).
+        fill_method : str, optional
+            If specified, fills nodata values using fill_nodata method.
+            Available methods are {'linear', 'nearest', 'cubic', 'rio_idw'}.
+
+        Returns
+        -------
+        list
+            Names of added model staticmap layers.
+        """
+        self.logger.info(f"Preparing staticmaps data from raster source {raster_fn}")
+        # Read raster data and select variables
+        ds = self.data_catalog.get_rasterdataset(
+            raster_fn,
+            geom=self.region,
+            buffer=2,
+            variables=variables,
+            single_var_as_array=False,
+        )
+        # Fill nodata
+        if fill_method is not None:
+            ds = ds.raster.interpolate_na(method=fill_method)
+        # Reprojection
+        ds_out = ds.raster.reproject_like(self.staticmaps, method=reproject_method)
+        # Add to staticmaps
+        self.set_staticmaps(ds_out)
+
+        # Update config
+        if wflow_variables is not None:
+            self.logger.info(
+                f"Updating the config for wflow_variables: {wflow_variables}"
+            )
+            if variables is None:
+                if len(ds_out.data_vars) == 1:
+                    variables = list(ds_out.data_vars.keys())
+                else:
+                    raise ValueError(
+                        f"Cannot update the toml if raster_fn has more than one variable and variables list is not provided."
+                    )
+
+            # Check on len
+            if len(wflow_variables) != len(variables):
+                raise ValueError(
+                    f"Length of variables {variables} do not match wflow_variables {wflow_variables}. Cannot update the toml."
+                )
+            else:
+                for i in range(len(variables)):
+                    self.set_config(wflow_variables[i], variables[i])
+
     def setup_precip_forcing(
         self,
         precip_fn: str = "era5",
@@ -1828,6 +1911,14 @@ class WflowModel(Model):
             ds = xr.open_dataset(fn, chunks={"time": 30}, decode_coords="all")
             for v in ds.data_vars:
                 self.set_forcing(ds[v])
+        elif "*" in str(fn):
+            self.logger.info(f"Read multiple forcing files using {fn}")
+            fns = list(fn.parent.glob(fn.name))
+            if len(fns) == 0:
+                raise IOError(f"No forcing files found using {fn}")
+            ds = xr.open_mfdataset(fns, chunks={"time": 30}, decode_coords="all")
+            for v in ds.data_vars:
+                self.set_forcing(ds[v])
 
     def write_forcing(
         self,
@@ -1857,7 +1948,7 @@ class WflowModel(Model):
             For more options, see https://pandas.pydata.org/pandas-docs/stable/user_guide/timeseries.html#offset-aliases
         chunksize: int, optional
             Chunksize on time dimension when saving to disk. By default 1.
-        decimals, int, optional
+        decimals: int, optional
             Round the ouput data to the given number of decimals.
         time_units: str, optional
             Common time units when writting several netcdf forcing files. By default "days since 1900-01-01T00:00:00".
@@ -1987,6 +2078,7 @@ class WflowModel(Model):
                 fns_out = os.path.relpath(fn_out, self.root)
                 fns_out = f"{str(fns_out)[0:-3]}_*.nc"
                 self.set_config("input.path_forcing", fns_out)
+                self.write_config()  # re-write config
                 for label, ds_gr in ds.resample(time=freq_out):
                     # ds_gr = group[1]
                     start = ds_gr["time"].dt.strftime("%Y%m%d")[0].item()
@@ -2151,7 +2243,6 @@ class WflowModel(Model):
         """Returns a river geometry as a geopandas.GeoDataFrame. If available, the
         stream order and upstream area values are added to the geometry properties.
         """
-        # import pdb; pdb.set_trace()
         if "rivers" in self.staticgeoms:
             gdf = self.staticgeoms["rivers"]
         elif self._MAPS["rivmsk"] in self.staticmaps: